requirements.txt
python-telegram-bot==13.15
pymongo==4.7.0
python-dotenv==1.0.0
google-generativeai==0.4.1
Pillow==10.0.0
